requirements.txt
gunicorn
eventlet
Flask==2.0.2
Flask-Cors==3.0.10
flask-sock==0.4.0
Flask-SocketIO==4.3.1
numpy==1.19.2
opencv-python-headless
Pillow
python-socketio==4.6.0
python-engineio==3.13.2
six==1.15.0
tensorflow-cpu==2.5.0
